⚡️ performance:  performance table
