+++ package.json
@@ -113,7 +113,7 @@
     "style-loader": "^2.0.0",
     "stylelint": "^13.0.0",
     "typescript": "^4.1.2",
-    "umi": "^3.2.10",
+    "umi": "^3.4.0",
     "umi-request": "^1.3.5",
     "umi-test": "^1.9.6",
     "umi-types": "^0.5.11",
