+++ packages/field/src/demos/base_test.tsx
@@ -184,21 +184,24 @@ export default () => {
             text="open"
             mode={state}
             valueType="select"
-            request={async () => [
-              { label: '全部', value: 'all' },
-              { label: '未解决', value: 'open' },
-              { label: '已解决', value: 'closed' },
-              { label: '解决中', value: 'processing' },
-              {
-                label: '特殊选项',
-                value: 'optGroup',
-                optionType: 'optGroup',
-                children: [
-                  { label: '不解决', value: 'no' },
-                  { label: '已废弃', value: 'clear' },
-                ],
-              },
-            ]}
+            request={async () => {
+              console.log('r');
+              return [
+                { label: '全部', value: 'all' },
+                { label: '未解决', value: 'open' },
+                { label: '已解决', value: 'closed' },
+                { label: '解决中', value: 'processing' },
+                {
+                  label: '特殊选项',
+                  value: 'optGroup',
+                  optionType: 'optGroup',
+                  children: [
+                    { label: '不解决', value: 'no' },
+                    { label: '已废弃', value: 'clear' },
+                  ],
+                },
+              ];
+            }}
           />
         </Descriptions.Item>
         <Descriptions.Item label="进度条">
