+++ packages/form/src/components/FieldSet/demos/components-other.tsx
@@ -71,12 +71,14 @@ const Demo = () => (
           fieldProps={{
             labelInValue: true,
           }}
-          request={async () => [
-            { label: '全部', value: 'all' },
-            { label: '未解决', value: 'open' },
-            { label: '已解决', value: 'closed' },
-            { label: '解决中', value: 'processing' },
-          ]}
+          request={async () => {
+            return [
+              { label: '全部', value: 'all' },
+              { label: '未解决', value: 'open' },
+              { label: '已解决', value: 'closed' },
+              { label: '解决中', value: 'processing' },
+            ];
+          }}
         />
         <ProFormSelect.SearchSelect
           name="userQuery"
