+++ packages/table/src/components/ColumnSetting/index.tsx
@@ -9,13 +9,14 @@ import {
 import type { TableColumnType } from 'antd';
 import { Checkbox, Tree, Popover, ConfigProvider, Tooltip } from 'antd';
 import classNames from 'classnames';
+import type { DataNode } from 'antd/lib/tree';
 
 import type { ColumnsState } from '../../container';
 import Container from '../../container';
-import type { DataNode } from 'antd/lib/tree';
-import './index.less';
-import { genColumnKey } from '../../utils';
+import { genColumnKey } from '../../utils/index';
 import type { ProColumns } from '../../typing';
+
+import './index.less';
 
 type ColumnSettingProps<T = any> = {
   columns: TableColumnType<T>[];
