+++ packages/table/src/demos/single.tsx
@@ -28,6 +28,7 @@ const columns: ProColumns<GithubIssueItem>[] = [
     width: 48,
   },
   {
+    width: 200,
     title: '标题',
     dataIndex: 'title',
     copyable: true,
